+++ src/client/.babelrc
@@ -1,0 +1,9 @@
+{
+  "presets": [
+    ["env", {
+      "targets": {
+        "browser": "latest 2 versions"
+      }
+    }]
+  ]
+}
