+++ test/server/requires.js
@@ -1,0 +1,3 @@
+require('babel-register')({
+  presets: [['env', {'targets': {'node': '6'}}]],
+});
